requirements.txt
prefect==2.14.18
pandas==2.3.1
SQLAlchemy==2.0.42
psycopg2-binary==2.9.10
python-dotenv==1.1.1
